+++ src/isla/z3_helpers.py
@@ -770,7 +770,9 @@ def smt_expr_to_str(  # noqa: C901
         assert len(qfd_var_stack) > idx
         return qfd_var_stack[idx]
     if z3.is_string_value(f):
-        return '"' + cast(str, f.as_string()).replace('"', r"\"") + '"'
+        result = '"' + cast(str, f.as_string()).replace('"', r"\"") + '"'
+        result = result.replace(r"\u{}", r"\u{0}")
+        return result
     if z3.is_int_value(f):
         return str(f.as_long())
     if z3.is_true(f):